Bug 1418072 - Move linkification code into its own module
