+++ tools/src/blame.rs
@@ -1,36 +1,13 @@
 use config;
+use links;
 
 use std::collections::BTreeMap;
 use rustc_serialize::json::Json;
-use regex::Regex;
 use git2;
 
 use chrono::naive::datetime::NaiveDateTime;
 use chrono::offset::fixed::FixedOffset;
 use chrono::datetime::DateTime;
-
-fn linkify_commit_header(s: &str) -> String {
-    lazy_static! {
-        static ref BUG_NUMBER_REGEX: Regex = {
-            Regex::new(r"\b(?P<bugno>[1-9][0-9]{4,9})\b").unwrap()
-        };
-        static ref SERVO_PR_REGEX: Regex = {
-            Regex::new(r"#(?P<prno>[1-9][0-9]*)\b").unwrap()
-        };
-    }
-    if s.starts_with("servo: ") {
-        SERVO_PR_REGEX.replace_all(s, "#<a href=\"https://github.com/servo/servo/pull/$prno\">$prno</a>")
-    } else {
-        BUG_NUMBER_REGEX.replace_all(s, "<a href=\"https://bugzilla.mozilla.org/show_bug.cgi?id=$bugno\">$bugno</a>")
-    }
-}
-
-#[test]
-fn test_linkify_servo_pr() {
-    let linkified =
-        linkify_commit_header("servo: Merge #1234 - stylo: Report a specific error for invalid CSS color values (from jdm:valueerr); r=heycam");
-    assert!(linkified.contains("github.com"), "{:?}", linkified);
-}
 
 pub fn commit_header(commit: &git2::Commit) -> Result<(String, String), &'static str> {
     fn entity_replace(s: &str) -> String {
@@ -41,7 +18,7 @@ pub fn commit_header(commit: &git2::Commit) -> Result<(String, String), &'static
     let mut iter = msg.split('\n');
     let header = iter.next().unwrap();
     let remainder = iter.collect::<Vec<_>>().join("\n");
-    let header = linkify_commit_header(&entity_replace(header));
+    let header = links::linkify_commit_header(&entity_replace(header));
     Ok((header, entity_replace(&remainder)))
 }
