+++ tools/src/lib.rs
@@ -14,6 +14,7 @@ pub mod output;
 pub mod languages;
 pub mod format;
 pub mod tokenize;
+pub mod links;
 
 pub fn find_source_file(path: &str, files_root: &str, objdir: &str) -> String {
     if path.starts_with("__GENERATED__") {
